# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Universal Cart API
  version: 0.2.0
  description: >-
    MVP API for authenticated universal cart import, matching, pricing comparison, checkout routing, audit trail, and
    privacy controls.
servers:
  - url: http://localhost:3001/api
    description: Local development
security:
  - bearerAuth: []
paths:
...
  /pricing/compare:
    post:
      summary: Compare source and destination effective totals
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - cartItemId
                - destinationRetailerProductId
              properties:
                cartItemId:
                  type: string
                destinationRetailerProductId:
                  type: string
      responses:
        '200':
          description: Pricing comparison
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PricingComparison'
...
  /pricing/{retailerProductId}/history:
    get:
      summary: Get retailer product price history and trend
      tags:
        - pricing
      responses:
        '200':
          description: Price history
        '400':
          $ref: '#/components/responses/Error'
      parameters:
        - name: retailerProductId
          in: path
          required: true
          schema:
            type: string
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
  schemas:
...
    PricingComparison:
      type: object
      properties:
        cartItemId:
          type: string
        source:
          type: object
          nullable: true
        destination:
          type: object
        recommendation:
          type: object
...